requirements.txt
openai
chromadb
pyyaml
requests
